Encapsulated logger into single module
